Fix FileTransfer upload for relative input path, Fix error message for non-json parsing of remote error, Add check for access token length, add cancellation token for recursive acl change and export fileproperties and add progress tracking to recursvie acl change, Fix overflow of offset in AdlsOutputStream.WriteAsync, Fix encoding of query parameters, Fix useragent to not contain localized values

+++ AdlsDotNetSDK/AclTools/AclProcessorStats.cs
@@ -9,27 +9,27 @@ namespace Microsoft.Azure.DataLake.Store.AclTools
         /// <summary>
         /// Number of files correctly processed
         /// </summary>
-        internal int FilesCorrect { get; set; }
+        internal long FilesCorrect { get; set; }
         /// <summary>
         /// Number of directories correctly processed
         /// </summary>
-        internal int DirectoriesCorrect { get; set; }
+        internal long DirectoriesCorrect { get; set; }
         /// <summary>
         /// Number of files processed
         /// </summary>
-        public int FilesProcessed { get; internal set; }
+        public long FilesProcessed { get; internal set; }
         /// <summary>
         /// Number of directories processed
         /// </summary>
-        public int DirectoryProcessed { get; internal set; }
+        public long DirectoryProcessed { get; internal set; }
 
-        public AclProcessorStats(int fileProcessed, int dirProcessed)
+        public AclProcessorStats(long fileProcessed, long dirProcessed)
         {
             FilesProcessed = fileProcessed;
             DirectoryProcessed = dirProcessed;
         }
 
-        internal AclProcessorStats(int fileProcessed, int dirProcessed,int fileCorrect,int dirCorrect):this(fileProcessed,dirProcessed)
+        internal AclProcessorStats(long fileProcessed, long dirProcessed, long fileCorrect, long dirCorrect):this(fileProcessed,dirProcessed)
         {
             FilesCorrect = fileCorrect;
             DirectoriesCorrect = dirCorrect;
